Moved BoxMetadataUpdateTask to be an internal object in BoxRequestsMetadata.UpdateFileMetadata
