+++ box-content-sdk/src/main/java/com/box/androidsdk/content/requests/BoxRequest.java
@@ -342,7 +342,6 @@ public abstract class BoxRequest<T extends BoxObject, R extends BoxRequest<T, R>
                 break;
         }
 
-        System.out.println(mStringBody);
         return mStringBody;
     }
